Separate actions/index.js and reducers/index.js into two files.
